fetch and parse ban lists in worker thread
