+++ modules/mod_tools.js
@@ -1,10 +1,7 @@
 "use strict";
 
-const axios = require('axios');
-const fs = require('fs');
-const path_resolve = require('path').resolve;
+const { Worker } = require('worker_threads');
 const bot_data = require('./data.js');
-const credentials = require('./credentials.js');
 const logger = require('./fletlog.js');
 const Fletalytics = require('./fletalytics');
 
@@ -64,46 +61,6 @@ function delay(t) {
     return new Promise(resolve => setTimeout(resolve, t));
 }
 
-/**
- * 
- * @param {Object} g_doc Google Document object
- * @returns {Array<string>} list of lines in doc
- */
-function get_doc_lines(g_doc) {
-    const doc_lines = [];
-    const doc_elements = g_doc.body.content;
-    for(const value of doc_elements) {
-        if('paragraph' in value) {
-            for(const p_elem of value.paragraph.elements) {
-                if(p_elem.textRun) {
-                    const raw_text = p_elem.textRun.content.trim();
-                    if(raw_text.length > 0) {
-                        doc_lines.push(raw_text);
-                    }
-                }
-            }
-        }
-    }
-    return doc_lines;
-}
-
-/**
- * 
- * @param {Object} g_sheet Google Spreadsheet objects
- * @returns {Array<string>} list of strings from sheet cells
- */
-function get_sheet_lines(g_sheet) {
-    const cell_list = [];
-    const sheet = g_sheet.sheets[0];
-    const rows = sheet.data[0].rowData;
-    for(const row of rows) {
-        if(row.values && row.values[0] && row.values[0].effectiveValue) {
-            cell_list.push(row.values[0].effectiveValue.stringValue);
-        }
-    }
-    return cell_list;
-}
-
 async function apply_bans(chat_client, full_ban_list, chat_cmd, delay_ms = 500) {
     for(const channel_name of chat_client.getChannels()) {
         if(!bot_data.is_bot_protected_channel(channel_name)) {
@@ -125,31 +82,6 @@ async function apply_bans(chat_client, full_ban_list, chat_cmd, delay_ms = 500) 
     }
 }
 
-async function fetch_doc(access_token, doc_data) {
-    let uri;
-    let parse_func;
-    switch (doc_data.ban_doc_type) {
-        case "doc":
-            uri = `https://docs.googleapis.com/v1/documents/${doc_data.ban_doc_id}?key=${credentials.get_google_key()}`;
-            parse_func = get_doc_lines;
-            break;
-        case "sheet":
-            uri = `https://sheets.googleapis.com/v4/spreadsheets/${doc_data.ban_doc_id}?key=${credentials.get_google_key()}&ranges=${doc_data.data_range}&includeGridData=true`;
-            parse_func = get_sheet_lines;
-            break;
-        default:
-            throw (`Unrecognized doc type ${doc_data.ban_doc_type}`);
-    }
-    const response = await axios({
-        url: uri,
-        method: 'get',
-        headers: {
-            'Authorization': `Bearer ${access_token}`
-        }
-    });
-    return parse_func(response.data);
-}
-
 /**
  * Pull list of banned usernames from Google Docs/Sheets, apply bans across relevant channels
  * @param {Object} chat_client tmi.js chat client
@@ -157,30 +89,35 @@ async function fetch_doc(access_token, doc_data) {
 function ban_wave(chat_client) {
     logger.log("Starting ban wave");
 
-    fs.readFile(mod_data_file, 'utf8', (error, str_data) => {
-        if(error) {
-            logger.error(error);
-            throw (error);
-        }
-        const mod_data = JSON.parse(str_data);
-        credentials.get_google_access_token()
-            .then((access_token) => {
-                const user_doc_requests = mod_data.user_ban_docs.map((doc_data) => fetch_doc(access_token, doc_data));
-                // TODO: once API for banning terms is exposed, update this to include blocking terms as part of ban wave
-                Promise.all(user_doc_requests)
-                    .then((doc_lines) => {
-                        const ban_list = Array.from(new Set(doc_lines.flat()));
-                        apply_bans(chat_client, ban_list, "/ban")
-                            .then(() => {
-                                logger.log("User ban wave completed");
-                            }).catch((err) => {
-                                logger.error(err);
-                            });
-                    }).catch((err) => {
-                        logger.error(err);
-                    });
+    const worker = new Worker('./workers/fetch_ban_list.js');
+    const fetch_promise = new Promise((resolve, reject) => {
+        worker.on('message', (message) => {
+            worker.terminate();
+            resolve(message);
+        });
+        worker.on('error', (err) => {
+            worker.terminate();
+            reject(err);
+        });
+        worker.on('exit', (exit_code) => {
+            if(exit_code != 0) {
+                reject(`Worker exited with code ${exit_code}`);
+            }
+        });
+    });
+    
+    fetch_promise.then((ban_lists) => {
+        apply_bans(chat_client, ban_lists.user_ban_list, "/ban")
+            .then(() => {
+                logger.log("User ban wave completed");
             }).catch((err) => {
                 logger.error(err);
             });
+        }).catch((err) => {
+            logger.error(err);
+        });
+
+    worker.postMessage({
+        mod_data_file
     });
 }
